+++ docetl/utils.py
@@ -8,19 +8,6 @@ from jinja2 import Environment, meta
 from litellm import completion_cost as lcc
 
 
-def render_jinja_template(template_string: str, data: Dict[str, Any]) -> str:
-    """
-    Render a Jinja2 template with the given data, ensuring protection against template injection vulnerabilities.
-    If the data is empty, return an empty string.
-    """
-    if not data:
-        return ""
-
-    env = Environment(autoescape=True)
-    template = env.from_string(template_string)
-    return template.render(input=data)
-
-
 def extract_jinja_variables(template_string: str) -> List[str]:
     """
     Extract variables from a Jinja2 template string.
